Cleaned up the table subclasses by using template design pattern. RemovePieceButtonCell Class will need to have the additional methods removed, yet I am going to wait till we actually start using it to remove those.
