+++ RepertoireListData/RepertoireList.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?><response><status><version>4.2</version><code>0</code><message>Success</message></status><songs><song><artist_name>Antoine Forqueray</artist_name><title>Harpsichord Suite No. 1 in D Minor: I La Laborde: Allemande</title></song><song><artist_name>Johann Jakob Froberger</artist_name><title>Toccata IX en Ut majeur</title></song><song><artist_name>Beethoven Quartet</artist_name><title>String Quartet No. 2 in A Major, Op. 68: IV. Theme &amp; Variations</title></song><song><artist_name>Quartetto Beethoven</artist_name><title>Allegro vivace</title></song><song><artist_name>Leopold Mozart</artist_name><title>6 Sonate per chiesa e da camera: No. 1 in F Major: III. Tempo di menuetto</title></song><song><artist_name>Mozart - Mozart-Ens. Amsterdam</artist_name><title>Symphony No. 25 - 1st mvt.</title></song><song><artist_name>Rameau</artist_name><title>El Tamborín</title></song><song><artist_name>Mozart Modern</artist_name><title>“Voi che sapete” from Le nozze di Figaro, K. 492</title></song><song><artist_name>Rameau</artist_name><title>Suite form Les fetes d'Hebe</title></song><song><artist_name>Johann Jakob Froberger</artist_name><title>Suite XVIII en sol mineur: II. Gigue</title></song><song><artist_name>Jean-Philippe Rameau</artist_name><title>Les fetes de Polymnie: Act I (La Fable) Scene 3: Les jours, pour nous, sont sans nuages (Une Suivante, Hebe)</title></song><song><artist_name>Alphonse Royer</artist_name><title>La favorita, Act II: Vien Leonora… De' nemici tuoi lo sdegno</title></song><song><artist_name>Rameau</artist_name><title>Elegie</title></song><song><artist_name>Mozart Symphonieorchester</artist_name><title>Aba Heidschi Bumbeidschi (Radio Version)</title></song><song><artist_name>Hardly Mozart</artist_name><title>About Last Night</title></song><song><artist_name>Kay</artist_name><title>M.A.J.O.R. (Gunslinger Remix)</title></song><song><artist_name>Rameau</artist_name><title>Le Temple de la Glorie: Suite</title></song><song><artist_name>Kay</artist_name><title>Away</title></song><song><artist_name>The Joe</artist_name><title>Meet Me On The Computer</title></song></songs></response>
+<?xml version="1.0" encoding="UTF-8" standalone="no"?><response><status><version>4.2</version><code>0</code><message>Success</message></status><songs><song><artist_name>Antoine Forqueray</artist_name><title>Harpsichord Suite No. 1 in D Minor: I La Laborde: Allemande</title></song><song><artist_name>Johann Jakob Froberger</artist_name><title>Toccata IX en Ut majeur</title></song><song><artist_name>Beethoven Quartet</artist_name><title>String Quartet No. 2 in A Major, Op. 68: IV. Theme &amp; Variations</title></song><song><artist_name>Quartetto Beethoven</artist_name><title>Allegro vivace</title></song><song><artist_name>Leopold Mozart</artist_name><title>6 Sonate per chiesa e da camera: No. 1 in F Major: III. Tempo di menuetto</title></song><song><artist_name>Mozart - Mozart-Ens. Amsterdam</artist_name><title>Symphony No. 25 - 1st mvt.</title></song><song><artist_name>Rameau</artist_name><title>El Tamborín</title></song><song><artist_name>Mozart Modern</artist_name><title>“Voi che sapete” from Le nozze di Figaro, K. 492</title></song><song><artist_name>Rameau</artist_name><title>Suite form Les fetes d'Hebe</title></song><song><artist_name>Johann Jakob Froberger</artist_name><title>Suite XVIII en sol mineur: II. Gigue</title></song><song><artist_name>Jean-Philippe Rameau</artist_name><title>Les fetes de Polymnie: Act I (La Fable) Scene 3: Les jours, pour nous, sont sans nuages (Une Suivante, Hebe)</title></song><song><artist_name>Alphonse Royer</artist_name><title>La favorita, Act II: Vien Leonora… De' nemici tuoi lo sdegno</title></song><song><artist_name>Rameau</artist_name><title>Elegie</title></song><song><artist_name>Mozart Symphonieorchester</artist_name><title>Aba Heidschi Bumbeidschi (Radio Version)</title></song><song><artist_name>Hardly Mozart</artist_name><title>About Last Night</title></song><song><artist_name>Kay</artist_name><title>M.A.J.O.R. (Gunslinger Remix)</title></song><song><artist_name>Rameau</artist_name><title>Le Temple de la Glorie: Suite</title></song><song><artist_name>Kay</artist_name><title>Away</title></song><song><artist_name>The Joe</artist_name><title>Meet Me On The Computer</title></song><song><artist_name>DJ John</artist_name><title>Que Calor (French Version) [feat. Dembow Zone]</title></song></songs></response>
